@@ -19,7 +19,6 @@ classifiers =
     License :: OSI Approved :: BSD License
     Natural Language :: English
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -34,7 +33,7 @@ install_requires =
     pyyaml
 # importlib; python_version == "2.6"
 packages = find:
-python_requires = >=3.6
+python_requires = >=3.7
 zip_safe = False
 
 [options.entry_points]
